[NEW] Adding PlayerItemService & Test
-In order to work, users must input their own steam key in PrivateKey.kt (temporary)
-Adding PrivateKey.kt & Android Manifest to .gitignore

+++ app/src/main/java/com/hdrussell/apps/steam/profile/services/GetPlayerItemParams.kt
@@ -2,6 +2,6 @@ package com.hdrussell.apps.steam.profile.services
 
 data class GetPlayerItemParams (
         val key: String,
-        val steamid: Double,
+        val steamid: Long,
         val format: String
 )
